Removed useless methods and clean code
